==> sources/R-Code/3_LA_Dev_Pattern.Rmd <==
---
title: "Development Outlook"
output: 
  html_document:
    css: "css/R_Knit.css"
    theme: cosmo
    code_folding: hide
    toc: false
knitr:
  opts_chunk:
    echo: FALSE
---

```{r setup, include=FALSE}
library(tidycensus)
library(tidyverse)
library(dplyr)
library(sf)
library(kableExtra)
library(ggplot2)
library(plotly)
library(readxl)
library(devtools)
library(readr)
library(leaflet)
library(leaflet.extras)
library(dbscan)
#devtools::install_github("GIScience/openrouteservice-r")
library(openrouteservice)
library(geosphere)
library(magrittr)
#remotes::install_github('rstudio/DT')
library(DT)

census_api_key("b3abcecc231fa30ccaa18cb5e854c30f1982fe3f", overwrite = TRUE, install = TRUE)
readRenviron("~/.Renviron")
```

```{r, warning=FALSE, message=FALSE, results='hide', include=FALSE}

LA_Permits <- read_excel("./data/excel/Building_and_Safety_Permits.xlsx")
CPA_Geo <- st_read("./data/shapefiles/Community_Plan_Areas_(CPA)/Community_Plan_Areas_(CPA).shp")
City_Geo <- st_read("./data/shapefiles/City_Boundary/City_Boundary.shp")
Dev_Permits <- LA_Permits %>% filter(PERMIT_TYPE == "Bldg-Addition" | PERMIT_TYPE == "Bldg-New" | PERMIT_TYPE == "Bldg-Demolition")


Use_Desc_Match <- read.csv("./Use_Desc_Match.csv") %>% select(Cat_1, Cat_2)
ZHVI_Raw <- read.csv("./ZHVI.csv")
Neighborhood <- st_read("./Neighborhood/Neighborhood_Demographics.shp") %>% st_transform(st_crs(CPA_Geo))
Neighborhood_Pop <- read.csv("./data/intermediate/Neighborhood_Pop.shp") %>% select(-X)
Neighborhood_Pop <- Neighborhood %>% select(name) %>% merge(Neighborhood_Pop, by='name')
```

```{r, eval=FALSE, include=FALSE}
# Color Palette
("#003954", "#16b7b1", "#eba5c7", "#9cb1c4", "#b3afc6","#D9E7F4")
 
("#468eca", "#e2a7c5", "#172d62", "#d1c47d") 

("#e7b4dd", "#d6cabc", "#70cfcb", "#067adf", "#01608e")


```

As a reminder, here are four tiers of factors that we look at:

Tier 1 (Finaled Permits or Certificate of Occupancy) | Tier 1.5 (Total and Issued Permits)  <br>
Tier 2 (Valuation & Area, from Permits) | Tier 3 (Population & Job Growth, Census) <br> <br>

```{r, include=F}
# tier 1
finaled_permits <- read.csv('./data/intermediate/Finaled_Permits.csv') # this is excluding 'Others'
finaled_permits <- finaled_permits %>% select(name, ERA, n) %>% pivot_wider(names_from=ERA, values_from=n)

tier_1 <- finaled_permits %>% mutate(
  permits_1 = case_when(`2010-2015`> median(`2010-2015`) ~ 1, TRUE ~ 0),
  permits_2 = case_when(`2016-2022`> median(`2016-2022`) ~ 1, TRUE ~ 0)
) %>%
  mutate(permits_status = case_when(
      permits_1 == 1 & permits_2 == 1 ~ "Constant Growth",
      permits_1 == 0 & permits_2 == 1 ~ "Recent Growth",
      permits_1 == 1 & permits_2 == 0 ~ "Stagnated Growth",
      TRUE ~ "Little or No Growth"
  ))


tier_1 <- Neighborhood_Pop %>%
  merge(tier_1, by="name") %>%
  mutate(permits_1_normalized = `2010-2015` / pop_16,
         permits_2_normalized = `2016-2022` / pop_22) %>%
  mutate(
  permits_1_normal = case_when(permits_1_normalized> median(permits_1_normalized) ~ 1, TRUE ~ 0),
  permits_2_normal = case_when(permits_2_normalized> median(permits_2_normalized) ~ 1, TRUE ~ 0)
) %>%
  mutate(permits_status_normalized = case_when(
      permits_1_normal == 1 & permits_2_normal == 1 ~ "Constant Growth",
      permits_1_normal == 0 & permits_2_normal == 1 ~ "Recent Growth",
      permits_1_normal == 1 & permits_2_normal == 0 ~ "Stagnated Growth",
      TRUE ~ "Little or No Growth"
  ))
```

```{r, include=F}
# tier 1.5
total_permits <- read.csv('./data/intermediate/Total_Permits.csv') 
issued_permits <- read.csv('./data/intermediate/issued_permits.csv') 

total_permits <- total_permits %>% select(name, ERA, n) %>% pivot_wider(names_from=ERA, values_from=n) %>% 
  rename(c('total_1_counts' = '2010-2015', 'total_2_counts' = '2016-2022'))
issued_permits <- issued_permits %>% select(name, ERA, n) %>% pivot_wider(names_from=ERA, values_from=n) %>%
    rename(c('issued_1_counts' = '2010-2015', 'issued_2_counts' = '2016-2022'))

tier_1.5 <- merge(total_permits, issued_permits, by='name')

tier_1.5 <- Neighborhood_Pop %>%
  merge(tier_1.5, by="name") %>%
  mutate(total_1_normal = total_1_counts / pop_16,
         total_2_normal = total_2_counts / pop_22,
         issued_1_normal = issued_1_counts / pop_16,
         issued_2_normal = issued_2_counts / pop_22) %>%
  mutate(
  total_1 = case_when(total_1_counts> median(total_1_counts, na.rm = TRUE) ~ 1, TRUE ~ 0),
  total_2 = case_when(total_2_counts> median(total_2_counts, na.rm = TRUE) ~ 1, TRUE ~ 0),
  total_1_normalized = case_when(total_1_normal> median(total_1_normal, na.rm = TRUE) ~ 1, TRUE ~ 0),
  total_2_normalized = case_when(total_2_normal> median(total_2_normal, na.rm = TRUE) ~ 1, TRUE ~ 0),
  issued_1 = case_when(issued_1_counts> median(issued_1_counts, na.rm = TRUE) ~ 1, TRUE ~ 0),
  issued_2 = case_when(issued_2_counts> median(issued_2_counts, na.rm = TRUE) ~ 1, TRUE ~ 0),
  issued_1_normalized = case_when(issued_1_normal> median(issued_1_normal, na.rm = TRUE) ~ 1, TRUE ~ 0),
  issued_2_normalized = case_when(issued_2_normal> median(issued_2_normal, na.rm = TRUE) ~ 1, TRUE ~ 0),
) %>%
  mutate(
    total_status = case_when(
      total_1 == 1 & total_2 == 1 ~ "Constant Growth",
      total_1 == 0 & total_2 == 1 ~ "Recent Growth",
      total_1 == 1 & total_2 == 0 ~ "Stagnated Growth",
      TRUE ~ "Little or No Growth"
    ),
    total_status_normalized = case_when(
      total_1_normalized == 1 & total_2_normalized == 1 ~ "Constant Growth",
      total_1_normalized == 0 & total_2_normalized == 1 ~ "Recent Growth",
      total_1_normalized == 1 & total_2_normalized == 0 ~ "Stagnated Growth",
      TRUE ~ "Little or No Growth"
    ),
    issue_status = case_when(
      issued_1 == 1 & issued_2 == 1 ~ "Constant Growth",
      issued_1 == 0 & issued_2 == 1 ~ "Recent Growth",
      issued_1 == 1 & issued_2 == 0 ~ "Stagnated Growth",
      TRUE ~ "Little or No Growth"
    ),
    issue_status_normalized = case_when(
      issued_1_normalized == 1 & issued_2_normalized == 1 ~ "Constant Growth",
      issued_1_normalized == 0 & issued_2_normalized == 1 ~ "Recent Growth",
      issued_1_normalized == 1 & issued_2_normalized == 0 ~ "Stagnated Growth",
      TRUE ~ "Little or No Growth"
    ))

```

```{r, include=F}
# Create map
tier_1 <- tier_1
pal_tier1 <- colorFactor(
  palette =  c("#003954", '#70cfcb',"#b3afc6", "#D9E7F4"),  
  levels = c("Constant Growth", "Recent Growth", "Stagnated Growth", "Little or No Growth")
)

Dev_map_tier1 <- leaflet() %>%
  addProviderTiles("CartoDB.Positron") %>%
  addPolygons(data=tier_1 %>% select(name, permits_status), 
              fillColor = ~pal_tier1(permits_status),
              fillOpacity = 1.0,
              color="#ffffff",
              weight = 1,
              label= ~paste(tier_1$name) %>%
                lapply(htmltools::HTML),
              group="Permits Status") %>%
    addPolygons(data=tier_1 %>% select(name, permits_status_normalized), 
              fillColor = ~pal_tier1(permits_status_normalized),
              fillOpacity = 1.0,
              color="#ffffff",
              weight = 1,
              label= ~paste(tier_1$name) %>%
                lapply(htmltools::HTML),
              group="Permits Status (Pop-Normalized)") %>%
  addLayersControl(overlayGroups = c("Permits Status", "Permits Status (Pop-Normalized)"),
                     options = layersControlOptions(collapsed = FALSE)) %>%
  hideGroup("Permits Status") %>%
  addLegend(pal = pal_tier1, 
            values = tier_1$permits_status,
            opacity = 0.7, 
            title = paste("Growth Status"),
            position = "bottomright",
           labels = c("Constant Growth", "Recent Growth", "Stagnated Growth", "Little or No Growth") %>%
   fitBounds(
    lng1 = -118.66819, lat1 = 33.70368,
    lng2 = -118.15537, lat2 = 34.33731
  ))

```

```{r, include=F}
# tier 2
area <- read.csv("./data/intermediate/Area.csv") %>% select(-X) %>% select(name, ERA, SquareFoot) %>% 
  mutate(SquareFoot = SquareFoot *1000000)
valuation <- read.csv("./data/intermediate/Valuation.csv") %>% select(-X) %>% mutate(Value = Value *1000000)

tier_2_area <- area %>% select(name, ERA, SquareFoot) %>%
  pivot_wider(names_from = ERA, values_from = SquareFoot) 

tier_2_area <- Neighborhood %>% select(name, Shape__Are) %>%
  merge(tier_2_area, by="name") %>%
  mutate(`2010-2015` = `2010-2015` / Shape__Are * 100, 
         `2016-2022` = `2016-2022` / Shape__Are * 100) %>%
  st_drop_geometry()# calculate the percentage of total neighborhood area that is being developed


tier_2_valuation <- valuation %>% select(name, ERA, Value) %>%
  pivot_wider(names_from = ERA, values_from = Value) 

tier_2_valuation <- Neighborhood %>% select(name) %>%
  merge(tier_2_valuation, by="name") %>%
  st_drop_geometry()
  
# Quantiles of 4 
quantile_area_1 <- quantile(tier_2_area$`2010-2015`, probs = seq(0, 1, length.out = 4))
quantile_area_2 <- quantile(tier_2_area$`2016-2022`, probs = seq(0, 1, length.out = 4))
quantile_value_1 <- quantile(tier_2_valuation$`2010-2015`, probs = seq(0, 1, length.out = 4))
quantile_value_2 <- quantile(tier_2_valuation$`2016-2022`, probs = seq(0, 1, length.out = 4))

# Categorize based on quantile data 
tier_2_area <- tier_2_area %>% mutate(
  area_1 = case_when(
    `2010-2015` <= quantile_area_1[2] ~ 0,
    `2010-2015` <= quantile_area_1[3] ~ 1,
    `2010-2015` <= quantile_area_1[4] ~ 2,
    TRUE ~ 3
  ),
  area_2 = case_when(
    `2016-2022` <= quantile_area_2[2] ~ 0,
    `2016-2022` <= quantile_area_2[3] ~ 1,
    `2016-2022` <= quantile_area_2[4] ~ 2,
    TRUE ~ 3
  )
) %>%
  mutate(area_both = area_1 + area_2) %>%
  mutate(
    Area_Status = case_when(
      area_1 >= 2 & area_2 >= 2 ~ "Constant Growth",
      area_1 < 2 & area_2 >= 2 ~ "Recent Growth",
      area_1 >= 2 & area_2 < 2 ~ "Stagnated Growth",
      TRUE ~ "Little or No Growth"
    )
  )


tier_2_valuation <- tier_2_valuation %>% mutate(
  value_1 = case_when(
    `2010-2015` <= quantile_value_1[2] ~ 0,
    `2010-2015` <= quantile_value_1[3] ~ 1,
    `2010-2015` <= quantile_value_1[4] ~ 2,
    TRUE ~ 3
  ),
  value_2 = case_when(
    `2016-2022` <= quantile_value_2[2] ~ 0,
    `2016-2022` <= quantile_value_2[3] ~ 1,
    `2016-2022` <= quantile_value_2[4] ~ 2,
    TRUE ~ 3
  )
) %>%
  mutate(value_both = value_1 + value_2) %>%
  mutate(
    Value_Status = case_when(
      value_1 >= 2 & value_2 >= 2 ~ "Constant Growth",
      value_1 < 2 & value_2 >= 2 ~ "Recent Growth",
      value_1 >= 2 & value_2 < 2 ~ "Stagnated Growth",
      TRUE ~ "Little or No Growth"
    )
  )


tier_2 <- Neighborhood %>% select(name, Shape__Are, Shape__Len) %>%
  merge(tier_2_area %>% select(-`2010-2015`, -`2016-2022`, -Shape__Are), by="name") 

tier_2 <- tier_2 %>%
  merge(tier_2_valuation %>% select(-`2010-2015`, -`2016-2022`), by="name") 

tier_2_geo <- st_as_sf(tier_2, geometry=geometry) 
```

```{r, include=F}
# tier 3
job_growth <- read.csv("./data/intermediate/Jobs.csv") %>% select(-X)
pop_growth <- read.csv("./data/intermediate/Pop_growth.csv") %>% select(-X)

tier_3_job <- job_growth %>% mutate(job_growth_1 = case_when(Change_10_16 >0 ~ 1, TRUE ~ 0),
                                    job_growth_2 = case_when(Change_16_21 >0 ~ 1, TRUE ~ 0),
                                    job_growth_both = case_when(job_growth_1 == 1 & job_growth_2 == 1 ~ 1, TRUE ~ 0))


tier_3_pop <- pop_growth %>% mutate(pop_growth_1 = case_when(Change_10_16 >0 ~ 1, TRUE ~ 0),
                                    pop_growth_2 = case_when(Change_16_22 >0 ~ 1, TRUE ~ 0),
                                    pop_growth_both = case_when(pop_growth_1 == 1 & pop_growth_2 == 1 ~ 1, TRUE ~ 0))

tier_3 <- Neighborhood %>% select(name, Shape__Are, Shape__Len) %>%
  merge(tier_3_job %>% select(-Change_10_16, -Change_16_21), by="name") %>%
  merge(tier_3_pop %>% select(-Change_10_16, -Change_16_22, -pop_10, -pop_16, -pop_22), by="name") 
  
tier_3 <- tier_3 %>% 
  mutate(Job_Status = case_when(
    job_growth_1 == 1 & job_growth_2 == 1 ~ "Constant Growth",
    job_growth_1 == 0 & job_growth_2 ==  1 ~ "Recent Growth",
    job_growth_1 == 1 & job_growth_2 == 0 ~ "Stagnated Growth",
    TRUE ~ "Little or No Growth"),
        Pop_Status = case_when(
    pop_growth_1 == 1 & pop_growth_2 == 1 ~ "Constant Growth",
    pop_growth_1 == 0 & pop_growth_2 == 1 ~ "Recent Growth",
    pop_growth_1 == 1 & pop_growth_2 == 0 ~ "Stagnated Growth",
    TRUE ~ "Little or No Growth")
  )


tier_3$Job_Status = factor(tier_3$Job_Status, levels = c("Constant Growth", "Recent Growth", "Stagnated Growth", "Little or No Growth" ))
tier_3$Pop_Status = factor(tier_3$Pop_Status, levels = c("Constant Growth", "Recent Growth", "Stagnated Growth", "Little or No Growth" ))


```

```{r, include=F}

# combining all tiers together 

combined_index <- tier_1 %>% st_drop_geometry() %>% select(name, permits_1_normal, permits_2_normal) %>% 
  merge(tier_1.5 %>% st_drop_geometry() %>% select(name, total_1_normalized, total_2_normalized, issued_1_normalized, issued_2_normalized), by='name') 
combined_index <- combined_index %>%
  merge(tier_2 %>% st_drop_geometry() %>% select(name, area_1, area_2, value_1, value_2), by='name')

combined_index <- tier_3 %>% select(name, job_growth_1, job_growth_2, pop_growth_1, pop_growth_2) %>%
  merge(combined_index , by='name')

# create new indexing based on tiers

combined_index <- combined_index %>% mutate(
  tier_1_1 = permits_1_normal, 
  tier_1_2 = permits_2_normal,
  tier_1.5_1 = permits_1_normal + total_1_normalized + issued_1_normalized, 
  tier_1.5_2 = permits_2_normal + total_2_normalized + issued_2_normalized, 
  tier_2_1 = permits_1_normal + total_1_normalized + issued_1_normalized + area_1 + value_1,
  tier_2_2 = permits_2_normal + total_2_normalized + issued_2_normalized + area_2 + value_2,
  tier_3_1 = permits_1_normal + total_1_normalized + issued_1_normalized + area_1 + value_1 + job_growth_1 + pop_growth_1,
  tier_3_2 = permits_2_normal + total_2_normalized + issued_2_normalized + area_2 + value_2 + job_growth_2 + pop_growth_2
)

combined_index <- combined_index %>% mutate(
  tier_1 = case_when(
      tier_1_1 == 1 & tier_1_2 == 1 ~ "Constant Growth",
      tier_1_1 == 0 & tier_1_2 == 1 ~ "Recent Growth",
      tier_1_1 == 1 & tier_1_2 == 0 ~ "Stagnated Growth",
      TRUE ~ "Little or No Growth"
  ),
  tier_1.5 = case_when(
      tier_1.5_1 >= median(tier_1.5_1) & tier_1.5_2 >= median(tier_1.5_2) ~ "Constant Growth",
      tier_1.5_1 < median(tier_1.5_1) & tier_1.5_2 >= median(tier_1.5_2) ~ "Recent Growth",
      tier_1.5_1 >= median(tier_1.5_1) & tier_1.5_2 < median(tier_1.5_2) ~ "Stagnated Growth",
      TRUE ~ "Little or No Growth"
  ),
  tier_2 = case_when(
      tier_2_1 >= median(tier_2_1) & tier_2_2 >= median(tier_2_2) ~ "Constant Growth",
      tier_2_1 < median(tier_2_1) & tier_2_2 >= median(tier_2_2) ~ "Recent Growth",
      tier_2_1 >= median(tier_2_1) & tier_2_2 < median(tier_2_2) ~ "Stagnated Growth",
      TRUE ~ "Little or No Growth"
  ),
  tier_3 = case_when(
      tier_3_1 >= median(tier_3_1) & tier_3_2 >= median(tier_3_2) ~ "Constant Growth",
      tier_3_1 < median(tier_3_1) & tier_3_2 >= median(tier_3_2) ~ "Recent Growth",
      tier_3_1 >= median(tier_3_1) & tier_3_2 < median(tier_3_2) ~ "Stagnated Growth",
      TRUE ~ "Little or No Growth"
  ),
)


combined_index <- combined_index %>% mutate(
  tier_b_1 = tier_1_1 + tier_1.5_1,
  tier_b_2 = tier_1_2 + tier_1.5_2,
  tier_c_1 = tier_1_1 + tier_1.5_1 + tier_2_1,
  tier_c_2 = tier_1_2 + tier_1.5_2 + tier_2_2,
  tier_d_1 = tier_1_1 + tier_1.5_1 + tier_2_1 + tier_3_1,
  tier_d_2 = tier_1_2 + tier_1.5_2 + tier_2_2 + tier_3_2
)

combined_index <- combined_index %>% mutate(
  tier_b = case_when(
      tier_b_1 >= median(tier_b_1) & tier_b_2 >= median(tier_b_2) ~ "Constant Growth",
      tier_b_1 < median(tier_b_1) & tier_b_2 >= median(tier_b_2) ~ "Recent Growth",
      tier_b_1 >= median(tier_b_1) & tier_b_2 < median(tier_b_2) ~ "Stagnated Growth",
      TRUE ~ "Little or No Growth"
  ),
  tier_c = case_when(
      tier_c_1 >= median(tier_c_1) & tier_c_2 >= median(tier_c_2) ~ "Constant Growth",
      tier_c_1 < median(tier_c_1) & tier_c_2 >= median(tier_c_2) ~ "Recent Growth",
      tier_c_1 >= median(tier_c_1) & tier_c_2 < median(tier_c_2) ~ "Stagnated Growth",
      TRUE ~ "Little or No Growth"
  ),
  tier_d = case_when(
      tier_d_1 >= median(tier_d_1) & tier_d_2 >= median(tier_d_2) ~ "Constant Growth",
      tier_d_1 < median(tier_d_1) & tier_d_2 >= median(tier_d_2) ~ "Recent Growth",
      tier_d_1 >= median(tier_d_1) & tier_d_2 < median(tier_d_2) ~ "Stagnated Growth",
      TRUE ~ "Little or No Growth"
  )
)


```

```{r, include=F}
growth <- combined_index %>% filter(tier_1 =="Constant Growth" & tier_2 =="Constant Growth" & tier_3 =="Constant Growth" & tier_1.5 =="Constant Growth" & tier_b =="Constant Growth" & tier_c =="Constant Growth" & tier_d =="Constant Growth" ) %>% select(name) %>% mutate(Growth = "Constant Growth")
no_growth <- combined_index %>% filter(tier_1 =="Little or No Growth" & tier_2 =="Little or No Growth" & tier_3 =="Little or No Growth" & tier_1.5 =="Little or No Growth"  & tier_b =="Little or No Growth"  & tier_c =="Little or No Growth"  & tier_d =="Little or No Growth") %>% select(name) %>% mutate(Growth = "Little or No Growth")

composite <- rbind(growth, no_growth)

```

```{r, include=F}

pal <- colorFactor(
  palette =  c("#003954", '#70cfcb',"#b3afc6", "#D9E7F4"),  
  levels = c("Constant Growth", "Recent Growth", "Stagnated Growth", "Little or No Growth")
)
Dev_map <- leaflet() %>%
  addProviderTiles("CartoDB.Positron") %>%
  addPolygons(data=combined_index %>% select(name, tier_1), 
              fillColor = ~pal(tier_1),
              fillOpacity = 1.0,
              color="#ffffff",
              weight = 1,
              label= ~paste(combined_index$name) %>%
                lapply(htmltools::HTML),
              group="Tier 1") %>%
    addPolygons(data=combined_index %>% select(name, tier_1.5), 
              fillColor = ~pal(tier_1.5),
              fillOpacity = 1.0,
              color="#ffffff",
              weight = 1,
              label= ~paste(combined_index$name) %>%
                lapply(htmltools::HTML),
              group="Tier 1.5") %>%
    addPolygons(data=combined_index %>% select(name, tier_2), 
              fillColor = ~pal(tier_2),
              fillOpacity = 1.0,
              color="#ffffff",
              weight = 1,
              label= ~paste(combined_index$name) %>%
                lapply(htmltools::HTML),
              group="Tier 2") %>%
    addPolygons(data=combined_index %>% select(name, tier_3), 
              fillColor = ~pal(tier_3),
              fillOpacity = 1.0,
              color="#ffffff",
              weight = 1,
              label= ~paste(combined_index$name) %>%
                lapply(htmltools::HTML),
              group="Tier 3") %>%
    addPolygons(data=combined_index %>% select(name, tier_b), 
              fillColor = ~pal(tier_b),
              fillOpacity = 1.0,
              color="#ffffff",
              weight = 1,
              label= ~paste(combined_index$name) %>%
                lapply(htmltools::HTML),
              group="Tier 1 + 1.5") %>%
  
    addPolygons(data=combined_index %>% select(name, tier_c), 
              fillColor = ~pal(tier_c),
              fillOpacity = 1.0,
              color="#ffffff",
              weight = 1,
              label= ~paste(combined_index$name) %>%
                lapply(htmltools::HTML),
              group="Tier 1 + 1.5 + 2") %>%
    addPolygons(data=combined_index %>% select(name, tier_d), 
              fillColor = ~pal(tier_d),
              fillOpacity = 1.0,
              color="#ffffff",
              weight = 1,
              label= ~paste(combined_index$name) %>%
                lapply(htmltools::HTML),
              group="Tier 1 + 1.5 + 2 + 3") %>%
      addPolygons(data=composite %>% select(name, Growth), 
              fillColor = ~pal(Growth),
              fillOpacity = 1.0,
              color="#ffffff",
              weight = 1,
              label= ~paste(composite$name) %>%
                lapply(htmltools::HTML),
              group="Composite") %>%
  addLayersControl(overlayGroups = c("Tier 1","Tier 1.5","Tier 2","Tier 3", "Tier 1 + 1.5", "Tier 1 + 1.5 + 2", "Tier 1 + 1.5 + 2 + 3", "Composite"),
                     options = layersControlOptions(collapsed = FALSE)) %>%
  hideGroup("Tier 1") %>%
  hideGroup("Tier 1.5") %>%
  hideGroup("Tier 2") %>%
  hideGroup("Tier 3") %>%
  hideGroup("Tier 1 + 1.5") %>%
  hideGroup("Tier 1 + 1.5 + 2") %>%
  hideGroup("Tier 1 + 1.5 + 2 + 3") %>%
  addLegend(pal = pal, 
            values = combined_index$tier_1,
            opacity = 0.7, 
            title = paste("Growth Status"),
            position = "bottomright",
           labels = c("Constant Growth", "Recent Growth", "Stagnated Growth", "Little or No Growth") %>%
   fitBounds(
    lng1 = -118.66819, lat1 = 33.70368,
    lng2 = -118.15537, lat2 = 34.33731
  ))
```

```{r, echo=F}
Dev_map
```

<br> This is the map showcasing development pattern in the City of LA in the past decade, which is one of the main goal for exploration of this study. Each layer represents single or multi-variate consideration when determining development growth. The map encompasses results when taking individual tiers or combination of tiers into considerations. While we see minor differences in spatial distribution of recent / stagnated growth when assessing with factors from different tiers, we see many neighborhoods in center and north areas that are categorized as constant growth across tiers while some towards the south have been experiencing less growth. <b>'Composite' layer</b> shows neighborhoods that are flagged as 'Constant Growth' or 'Little or No Growth' across / regardless of which tiers are considered for assessments. Following table lists all neighborhoods that fall in those two categories. 
<br>
<br>
<b>It is important to note that permit counts are normalized by population in each neighborhood, and area squarefootage is normalized by total area of the neighborhood. </b><br>
Turn on and off layers to compare different results when using each tier of indicators or combined -- they are especially helpful for visualizing areas with more recent growth and stagnated growth. 

<br>
<br>


```{r, echo=F}
composite <- composite %>% st_drop_geometry() %>% select(name, Growth)
datatable(composite)
```

<br> <br>

The result of this study is helpful for visualizing development's spatial patterns for the period of 2010-2022, which may help channel more resources to certain neighborhoods or areas. Together with factors such as median household income (inflation adjusted) and policy-level decision and implementations such as grants, we can have a better understanding of neighborhoods that are experiencing little development or stagnated growth since 2015, where resources should be allocated more towards. 

```{r, echo=F, include=F}
Dev_map_tier1
```


```{r, include=F}
# Create map

pal_tier1.5 <- colorFactor(
  palette =  c("#003954", '#70cfcb',"#b3afc6", "#D9E7F4"),  
  levels = c("Constant Growth", "Recent Growth", "Stagnated Growth", "Little or No Growth")
)

bbox <- st_bbox(City_Geo)   
Dev_map_tier1.5 <- leaflet() %>%
  addProviderTiles("CartoDB.Positron") %>%
  addPolygons(data=tier_1.5 %>% select(name, total_status), 
              fillColor = ~pal_tier1.5(total_status),
              fillOpacity = 1.0,
              color="#ffffff",
              weight = 1,
              label= ~paste(tier_1.5$name) %>%
                lapply(htmltools::HTML),
              group="Total Status") %>%
    addPolygons(data=tier_1.5 %>% select(name, total_status_normalized), 
              fillColor = ~pal_tier1.5(total_status_normalized),
              fillOpacity = 1.0,
              color="#ffffff",
              weight = 1,
              label= ~paste(tier_1.5$name) %>%
                lapply(htmltools::HTML),
              group="Total Status (Pop-Normalized)") %>%
    addPolygons(data=tier_1.5 %>% select(name, issue_status), 
              fillColor = ~pal_tier1.5(issue_status),
              fillOpacity = 1.0,
              color="#ffffff",
              weight = 1,
              label= ~paste(tier_1.5$name) %>%
                lapply(htmltools::HTML),
              group="Issued Status") %>%
    addPolygons(data=tier_1.5 %>% select(name, issue_status_normalized), 
              fillColor = ~pal_tier1.5(issue_status_normalized),
              fillOpacity = 1.0,
              color="#ffffff",
              weight = 1,
              label= ~paste(tier_1.5$name) %>%
                lapply(htmltools::HTML),
              group="Issued Status (Pop-Normalized)") %>%
  addLayersControl(overlayGroups = c("Total Status", "Total Status (Pop-Normalized)", "Issued Status", "Issued Status (Pop-Normalized)"),
                     options = layersControlOptions(collapsed = FALSE)) %>%
  hideGroup("Total Status") %>%
  hideGroup("Total Status (Pop-Normalized)") %>%
  hideGroup("Issued Status") %>%
  addLegend(pal = pal_tier1.5, 
            values = tier_1.5$issue_status,
            opacity = 0.7, 
            title = paste("Growth Status"),
            position = "bottomright",
           labels = c("Constant Growth", "Recent Growth", "Stagnated Growth", "Little or No Growth") %>%
   fitBounds(
    lng1 = -118.66819, lat1 = 33.70368,
    lng2 = -118.15537, lat2 = 34.33731
  ))

```

```{r, echo=FALSE, include=F}
Dev_map_tier1.5
```

```{r, include=F}

pal_tier2 <- colorFactor(
  palette =  c("#003954", '#70cfcb',"#b3afc6", "#D9E7F4"),  
  levels = c("Constant Growth", "Recent Growth", "Stagnated Growth", "Little or No Growth")
)

bbox <- st_bbox(City_Geo)   
Dev_map_tier2 <- leaflet() %>%
  addProviderTiles("CartoDB.Positron") %>%
  addPolygons(data=tier_2_geo %>% select(name, Area_Status), 
              fillColor = ~pal_tier2(Area_Status),
              fillOpacity = 1.0,
              color="#ffffff",
              weight = 1,
              label= ~paste(tier_2_geo$name) %>%
                lapply(htmltools::HTML),
              group="Area Status") %>%
  addPolygons(data=tier_2_geo %>% select(name, Value_Status), 
              fillColor = ~pal_tier2(Value_Status),
              fillOpacity = 1.0,
              color="#ffffff",
              weight=1, 
              label= paste(tier_2_geo$name) %>%
                lapply(htmltools::HTML), 
              group="Value Status") %>%
  addLegend(pal = pal_tier2, 
            values = tier_2_geo$Value_Status,
            opacity = 0.7, 
            title = paste("Growth Status"),
            position = "bottomright",
           labels = c("Constant Growth", "Recent Growth", "Stagnated Growth", "Little or No Growth")
 ) %>%
  addLayersControl(overlayGroups = c("Area Status", "Value Status"),
                     options = layersControlOptions(collapsed = FALSE)) %>%
  hideGroup("Area Status") %>%
   fitBounds(
    lng1 = -118.66819, lat1 = 33.70368,
    lng2 = -118.15537, lat2 = 34.33731
  )

Dev_map_tier2
```


```{r, echo=F, include=F}

pal_tier3 <- colorFactor(
  palette =  c("#003954", '#70cfcb',"#b3afc6", "#D9E7F4"),  
  levels = c("Constant Growth", "Recent Growth", "Stagnated Growth", "Little or No Growth")
)

bbox <- st_bbox(City_Geo)   
Dev_map_tier3 <- leaflet() %>%
  addProviderTiles("CartoDB.Positron") %>%
  addPolygons(data=tier_3 %>% select(name, Job_Status), 
              fillColor = ~pal_tier3(Job_Status),
              fillOpacity = 1.0,
              color="#ffffff",
              weight = 1,
              label= ~paste(tier_3$name) %>%
                lapply(htmltools::HTML),
              group="Job Status") %>%
  addPolygons(data=tier_3 %>% select(name, Pop_Status), 
              fillColor = ~pal_tier3(Pop_Status),
              fillOpacity = 1.0,
              color="#ffffff",
              weight=1, 
              label= paste(tier_3$name) %>%
                lapply(htmltools::HTML), 
              group="Pop Status") %>%
  addLegend(pal = pal_tier3, 
            values = tier_3$Pop_Status,
            opacity = 0.7, 
            title = paste("Growth Status"),
            position = "bottomright",
           labels = c("Constant Growth", "Recent Growth", "Stagnated Growth", "Little or No Growth")
 ) %>%
  addLayersControl(overlayGroups = c("Job Status", "Pop Status"),
                     options = layersControlOptions(collapsed = FALSE)) %>%
  hideGroup("Job Status") %>%
   fitBounds(
    lng1 = -118.66819, lat1 = 33.70368,
    lng2 = -118.15537, lat2 = 34.33731
  )

Dev_map_tier3
```

<br> <br> <br>
